+++ porko-service/src/main/java/io/porko/widget/domain/OrderedMemberWidgets.java
@@ -1,5 +1,8 @@
 package io.porko.widget.domain;
 
+import static io.porko.widget.exception.WidgetErrorCode.DUPLICATED_SEQUENCE;
+import static io.porko.widget.exception.WidgetErrorCode.INVALID_REQUEST_ORDERED_WIDGET_COUNT;
+
 import io.porko.member.domain.Member;
 import io.porko.widget.controller.model.ModifyMemberWidgetOrderDto;
 import io.porko.widget.controller.model.ReorderWidgetRequest;
@@ -8,16 +11,20 @@ import io.porko.widget.exception.WidgetErrorCode;
 import io.porko.widget.exception.WidgetException;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 public record OrderedMemberWidgets(
     List<MemberWidget> elements
 ) {
+    public static final int ORDERED_WIDGET_COUNT = 6;
+
     public static OrderedMemberWidgets of(
         Member member,
         WidgetsResponse widgetsResponse,
         ReorderWidgetRequest reorderWidgetRequest
     ) {
+        validateRequest(reorderWidgetRequest);
         Map<Long, Widget> widgetMap = toMap(widgetsResponse.toWidgets());
         List<MemberWidget> memberWidgets = reorderWidgetRequest.elements().stream()
             .map(it -> processSequencedWidget(member, it, widgetMap))
@@ -25,6 +32,32 @@ public record OrderedMemberWidgets(
         widgetMap.forEach((key, value) -> memberWidgets.add(processUnsequencedWidget(member, value)));
 
         return new OrderedMemberWidgets(memberWidgets);
+    }
+
+    private static void validateRequest(ReorderWidgetRequest reorderWidgetRequest) {
+        List<ModifyMemberWidgetOrderDto> elements = reorderWidgetRequest.elements();
+        validateOrderedWidgetCount(elements);
+        validateSequenceRange(elements);
+    }
+
+    private static void validateOrderedWidgetCount(List<ModifyMemberWidgetOrderDto> elements) {
+        if (elements.size() != ORDERED_WIDGET_COUNT) {
+            throw new WidgetException(INVALID_REQUEST_ORDERED_WIDGET_COUNT);
+        }
+    }
+
+    private static void validateSequenceRange(List<ModifyMemberWidgetOrderDto> elements) {
+        Set<Integer> sequenceSet = elements.stream()
+            .map(ModifyMemberWidgetOrderDto::sequence)
+            .collect(Collectors.toSet());
+
+        if (hasDuplicateSequence(sequenceSet)) {
+            throw new WidgetException(DUPLICATED_SEQUENCE, elements);
+        }
+    }
+
+    private static boolean hasDuplicateSequence(Set<Integer> sequenceSet) {
+        return sequenceSet.size() != ORDERED_WIDGET_COUNT;
     }
 
     private static MemberWidget processSequencedWidget(Member member, ModifyMemberWidgetOrderDto it, Map<Long, Widget> widgetMap) {
